+++ CHANGELOG.md
@@ -1,8 +1,13 @@
+## 2.2.0
+
+* Added `inList` and `inCollection` filters.
+* Updated readme and example.
+
 ## 2.1.0
 
 * Added `toEncodedString()` and normal `toString()` does not encode the query.
-* Fixed typo in `scr` -> `src` folder
-* Example for nested odata queries
+* Fixed typo in `scr` -> `src` folder.
+* Example for nested odata queries.
 
 ## 2.0.0